refactor: move epoch-change handling from node.rs into node/epoch_change module
Mechanical move with no logic changes: the EpochChangeEvent dispatch,
the EpochChangeStart/EpochChangeDone handlers and their descendants,
BeginCommitteeChangeAction, and enter_recovery_mode now live in
crates/walrus-service/src/node/epoch_change.rs.
